+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
+### Added
+
+- Type Safe CSS URL Handler will be force-installed if the JVM does not pick it up
+
+
 ## [1.4.7] - 2016-05-10
 
 ### Added